@@ -1,323 +1,85 @@
-part of auto_size_text;
+import 'package:auto_size_text/src/auto_size_builder/auto_size_builder.dart';
+import 'package:flutter/widgets.dart';
 
 /// Flutter widget that automatically resizes text to fit perfectly within its
 /// bounds.
 ///
 /// All size constraints as well as maxLines are taken into account. If the text
 /// overflows anyway, you should check if the parent widget actually constraints
-/// the size of this widget.
-class AutoSizeText extends StatefulWidget {
-  /// Creates a [AutoSizeText] widget.
-  ///
-  /// If the [style] argument is null, the text will use the style from the
-  /// closest enclosing [DefaultTextStyle].
-  const AutoSizeText(
-    String this.data, {
-    Key? key,
-    this.textKey,
-    this.style,
-    this.strutStyle,
-    this.minFontSize = 12,
-    this.maxFontSize = double.infinity,
-    this.stepGranularity = 1,
-    this.presetFontSizes,
-    this.group,
-    this.textAlign,
-    this.textDirection,
-    this.locale,
-    this.softWrap,
-    this.wrapWords = true,
-    this.overflow,
-    this.overflowReplacement,
-    this.textScaleFactor,
-    this.maxLines,
-    this.semanticsLabel,
-  })  : textSpan = null,
-        autofocus = false,
-        showCursor = false,
-        cursorWidth = 2.0,
-        cursorHeight = null,
-        cursorRadius = null,
-        focusNode = null,
-        cursorColor = null,
-        enableInteractiveSelection = false,
-        selectionControls = null,
-        dragStartBehavior = DragStartBehavior.start,
-        toolbarOptions = null,
-        onTap = null,
-        scrollPhysics = null,
-        onSelectionChanged = null,
-        _isSelectableText = false,
-        super(key: key);
-
-  /// Creates a [AutoSizeText] widget with a [TextSpan].
-  const AutoSizeText.rich(
-    TextSpan this.textSpan, {
-    Key? key,
-    this.textKey,
-    this.style,
-    this.strutStyle,
-    this.minFontSize = 12,
-    this.maxFontSize = double.infinity,
-    this.stepGranularity = 1,
-    this.presetFontSizes,
-    this.group,
-    this.textAlign,
-    this.textDirection,
-    this.locale,
-    this.softWrap,
-    this.wrapWords = true,
-    this.overflow,
-    this.overflowReplacement,
-    this.textScaleFactor,
-    this.maxLines,
-    this.semanticsLabel,
-  })  : data = null,
-        autofocus = false,
-        showCursor = false,
-        cursorWidth = 2.0,
-        cursorHeight = null,
-        cursorRadius = null,
-        focusNode = null,
-        cursorColor = null,
-        enableInteractiveSelection = false,
-        selectionControls = null,
-        dragStartBehavior = DragStartBehavior.start,
-        toolbarOptions = null,
-        onTap = null,
-        scrollPhysics = null,
-        onSelectionChanged = null,
-        _isSelectableText = false,
-        super(key: key);
-
-  /// Creates a selectable [AutoSizeText] widget with a [TextSpan].
-  const AutoSizeText.richSelectable(
-    this.textSpan, {
-    Key? key,
-    this.textKey,
-    this.style,
-    this.strutStyle,
-    this.minFontSize = 12,
-    this.maxFontSize = double.infinity,
-    this.stepGranularity = 1,
-    this.presetFontSizes,
-    this.group,
-    this.textAlign,
-    this.textDirection,
-    this.wrapWords = true,
-    this.textScaleFactor,
-    this.maxLines,
-    this.autofocus = false,
-    this.showCursor = false,
-    this.cursorWidth = 2.0,
-    this.cursorHeight,
-    this.cursorRadius,
-    this.focusNode,
-    this.cursorColor,
-    this.enableInteractiveSelection = true,
-    this.selectionControls,
-    this.dragStartBehavior = DragStartBehavior.start,
-    this.toolbarOptions,
-    this.onTap,
-    this.scrollPhysics,
-    this.onSelectionChanged,
-  })  : assert(textSpan != null,
-            'A non-null TextSpan must be provided to a AutoSizeText.rich widget.'),
-        data = null,
-        locale = null,
-        softWrap = null,
-        overflow = null,
-        overflowReplacement = null,
-        semanticsLabel = null,
-        _isSelectableText = true,
-        super(key: key);
-
-  /// Creates a selectable [AutoSizeText] widget.
-  ///
-  /// If the [style] argument is null, the text will use the style from the
-  /// closest enclosing [DefaultTextStyle].
-  const AutoSizeText.selectable(
-    this.data, {
-    Key? key,
-    this.textKey,
-    this.style,
-    this.strutStyle,
-    this.minFontSize = 12,
-    this.maxFontSize = double.infinity,
-    this.stepGranularity = 1,
-    this.presetFontSizes,
-    this.group,
-    this.textAlign,
-    this.textDirection,
-    this.wrapWords = true,
-    this.textScaleFactor,
-    this.maxLines,
-    this.autofocus = false,
-    this.showCursor = false,
-    this.cursorWidth = 2.0,
-    this.cursorHeight,
-    this.cursorRadius,
-    this.focusNode,
-    this.cursorColor,
-    this.enableInteractiveSelection = true,
-    this.selectionControls,
-    this.dragStartBehavior = DragStartBehavior.start,
-    this.toolbarOptions,
-    this.onTap,
-    this.scrollPhysics,
-    this.onSelectionChanged,
-  })  : assert(data != null,
-            'A non-null String must be provided to a AutoSizeText widget.'),
-        textSpan = null,
-        locale = null,
-        softWrap = null,
-        overflow = null,
-        overflowReplacement = null,
-        semanticsLabel = null,
-        _isSelectableText = true,
-        super(key: key);
-
-  /// Sets the key for the resulting [Text] widget.
+/// the size of this
+class AutoSizeText extends StatelessWidget {
+  /// {@template auto_size_text.textKey}
+  /// Sets the key for the resulting [Text]
   ///
   /// This allows you to find the actual `Text` widget built by `AutoSizeText`.
+  /// It is useful if you want to be able to find the widget to modify it in
+  /// the [State.build] method of the [StatefulWidget] that `AutoSizeText` is
+  /// inserted into.
+  /// {@endtemplate}
   final Key? textKey;
 
-  /// The text to display.
-  ///
-  /// This will be null if a [textSpan] is provided instead.
+  /// {@macro flutter.painting.textPainter.textWidthBasis}
   final String? data;
 
+  /// {@template auto_size_text.textSpan}
   /// The text to display as a [TextSpan].
   ///
   /// This will be null if [data] is provided instead.
+  /// {@endtemplate}
   final TextSpan? textSpan;
 
+  /// {@template auto_size_text.style}
   /// If non-null, the style to use for this text.
   ///
   /// If the style's "inherit" property is true, the style will be merged with
   /// the closest enclosing [DefaultTextStyle]. Otherwise, the style will
   /// replace the closest enclosing [DefaultTextStyle].
+  /// {@endtemplate}
   final TextStyle? style;
 
-  // The default font size if none is specified.
-  static const double _defaultFontSize = 14;
-
-  /// The strut style to use. Strut style defines the strut, which sets minimum
-  /// vertical layout metrics.
-  ///
-  /// Omitting or providing null will disable strut.
-  ///
-  /// Omitting or providing null for any properties of [StrutStyle] will result
-  /// in default values being used. It is highly recommended to at least specify
-  /// a font size.
-  ///
-  /// See [StrutStyle] for details.
+  /// {@macro flutter.painting.textPainter.strutStyle}
   final StrutStyle? strutStyle;
 
-  /// The minimum text size constraint to be used when auto-sizing text.
-  ///
-  /// Is being ignored if [presetFontSizes] is set.
-  final double minFontSize;
-
-  /// The maximum text size constraint to be used when auto-sizing text.
-  ///
-  /// Is being ignored if [presetFontSizes] is set.
-  final double maxFontSize;
-
-  /// The step size in which the font size is being adapted to constraints.
-  ///
-  /// The Text scales uniformly in a range between [minFontSize] and
-  /// [maxFontSize].
-  /// Each increment occurs as per the step size set in stepGranularity.
-  ///
-  /// Most of the time you don't want a stepGranularity below 1.0.
-  ///
-  /// Is being ignored if [presetFontSizes] is set.
-  final double stepGranularity;
-
-  /// Predefines all the possible font sizes.
-  ///
-  /// **Important:** PresetFontSizes have to be in descending order.
-  final List<double>? presetFontSizes;
-
-  /// Synchronizes the size of multiple [AutoSizeText]s.
-  ///
-  /// If you want multiple [AutoSizeText]s to have the same text size, give all
-  /// of them the same [AutoSizeGroup] instance. All of them will have the
-  /// size of the smallest [AutoSizeText]
-  final AutoSizeGroup? group;
-
-  /// How the text should be aligned horizontally.
+  /// {@macro flutter.widgets.editableText.textAlign}
   final TextAlign? textAlign;
 
-  /// The directionality of the text.
-  ///
-  /// This decides how [textAlign] values like [TextAlign.start] and
-  /// [TextAlign.end] are interpreted.
-  ///
-  /// This is also used to disambiguate how to render bidirectional text. For
-  /// example, if the [data] is an English phrase followed by a Hebrew phrase,
-  /// in a [TextDirection.ltr] context the English phrase will be on the left
-  /// and the Hebrew phrase to its right, while in a [TextDirection.rtl]
-  /// context, the English phrase will be on the right and the Hebrew phrase on
-  /// its left.
-  ///
-  /// Defaults to the ambient [Directionality], if any.
+  /// {@macro flutter.widgets.editableText.textDirection}
   final TextDirection? textDirection;
 
+  /// {@template auto_size_text.locale}
   /// Used to select a font when the same Unicode character can
   /// be rendered differently, depending on the locale.
   ///
   /// It's rarely necessary to set this property. By default its value
   /// is inherited from the enclosing app with `Localizations.localeOf(context)`.
+  ///
+  /// See [RenderParagraph.locale] for more information.
+  /// {@endtemplate}
   final Locale? locale;
 
+  /// {@template auto_size_text.softWrap}
   /// Whether the text should break at soft line breaks.
   ///
   /// If false, the glyphs in the text will be positioned as if there was
   /// unlimited horizontal space.
+  /// {@endtemplate}
   final bool? softWrap;
 
-  /// Whether words which don't fit in one line should be wrapped.
-  ///
-  /// If false, the fontSize is lowered as far as possible until all words fit
-  /// into a single line.
-  final bool wrapWords;
-
+  /// {@template auto_size_text.overflow}
   /// How visual overflow should be handled.
   ///
   /// Defaults to retrieving the value from the nearest [DefaultTextStyle] ancestor.
+  /// If there is no ancestor, [TextOverflow.clip] is used.
+  /// {@endtemplate}
   final TextOverflow? overflow;
 
-  /// If the text is overflowing and does not fit its bounds, this widget is
-  /// displayed instead.
-  final Widget? overflowReplacement;
-
-  /// The number of font pixels for each logical pixel.
-  ///
-  /// For example, if the text scale factor is 1.5, text will be 50% larger than
-  /// the specified font size.
-  ///
-  /// This property also affects [minFontSize], [maxFontSize] and [presetFontSizes].
-  ///
-  /// The value given to the constructor as textScaleFactor. If null, will
-  /// use the [MediaQueryData.textScaleFactor] obtained from the ambient
-  /// [MediaQuery], or 1.0 if there is no [MediaQuery] in scope.
+  /// {@macro flutter.widgets.editableText.textScaleFactor}
   final double? textScaleFactor;
 
-  /// An optional maximum number of lines for the text to span, wrapping if necessary.
-  /// If the text exceeds the given number of lines, it will be resized according
-  /// to the specified bounds and if necessary truncated according to [overflow].
-  ///
-  /// If this is 1, text will not wrap. Otherwise, text will be wrapped at the
-  /// edge of the box.
-  ///
-  /// If this is null, but there is an ambient [DefaultTextStyle] that specifies
-  /// an explicit number for its [DefaultTextStyle.maxLines], then the
-  /// [DefaultTextStyle] value will take precedence. You can use a [RichText]
-  /// widget directly to entirely override the [DefaultTextStyle].
+  /// {@macro flutter.widgets.editableText.maxLines}
   final int? maxLines;
 
+  /// {@template auto_size_text.semanticsLabel}
   /// An alternative semantics label for this text.
   ///
   /// If present, the semantics of this widget will contain this value instead
@@ -330,395 +92,171 @@ class AutoSizeText extends StatefulWidget {
   /// ```dart
   /// AutoSizeText(r'$$', semanticsLabel: 'Double dollars')
   /// ```
+  /// {@endtemplate}
   final String? semanticsLabel;
 
-  final bool _isSelectableText;
+  /// {@macro flutter.dart:ui.textHeightBehavior}
+  final TextHeightBehavior? textHeightBehavior;
 
-  /// {@macro flutter.widgets.editableText.autofocus}
-  final bool autofocus;
+  /// {@macro flutter.painting.textPainter.textWidthBasis}
+  final TextWidthBasis? textWidthBasis;
 
-  /// The color to use when painting the cursor.
+  /// {@template auto_size_text.minFontSize}
+  /// The minimum text size constraint to be used when auto-sizing text.
   ///
-  /// Defaults to the theme's `cursorColor` when null.
-  final Color? cursorColor;
+  /// Is being ignored if [presetFontSizes] is set.
+  /// {@endtemplate}
+  final double? minFontSize;
 
-  /// {@macro flutter.widgets.editableText.cursorHeight}
-  final double? cursorHeight;
+  /// {@template auto_size_text.maxFontSize}
+  /// The maximum text size constraint to be used when auto-sizing text.
+  ///
+  /// Is being ignored if [presetFontSizes] is set.
+  /// {@endtemplate}
+  final double? maxFontSize;
 
-  /// {@macro flutter.widgets.editableText.cursorRadius}
-  final Radius? cursorRadius;
+  /// {@template auto_size_text.stepGranularity}
+  /// The step size in which the font size is being adapted to constraints.
+  ///
+  /// The Text scales uniformly in a range between [minFontSize] and
+  /// [maxFontSize].
+  /// Each increment occurs as per the step size set in stepGranularity.
+  ///
+  /// Most of the time you don't want a stepGranularity below 1.0.
+  ///
+  /// Is being ignored if [presetFontSizes] is set.
+  /// {@endtemplate}
+  final double? stepGranularity;
 
-  /// {@macro flutter.widgets.editableText.cursorWidth}
-  final double cursorWidth;
+  /// {@template auto_size_text.presetFontSizes}
+  /// Predefines all the possible font sizes.
+  ///
+  /// **Important:** PresetFontSizes have to be in descending order.
+  /// {@endtemplate}
+  final List<double>? presetFontSizes;
 
-  /// {@macro flutter.widgets.scrollable.dragStartBehavior}
-  final DragStartBehavior dragStartBehavior;
+  /// Synchronizes the size of multiple [AutoSizeText]s.
+  ///
+  /// If you want multiple [AutoSizeText]s to have the same text size, give all
+  /// of them the same [AutoSizeGroup] instance. All of them will have the
+  /// size of the smallest [AutoSizeText]
+  //final AutoSizeGroup? group;
 
-  /// Whether to enable user interface affordances for changing the text selection.
+  /// {@template auto_size_text.wrapWords}
+  /// Whether words which don't fit in one line should be wrapped.
   ///
-  /// For example, setting this to true will enable features such as long-pressing the TextField to select text and show the cut/copy/paste menu, and tapping to move the text caret.
-  ///
-  /// When this is false, the text selection cannot be adjusted by the user, text cannot be copied, and the user cannot paste into the text field from the clipboard.
-  final bool enableInteractiveSelection;
+  /// If false, the fontSize is lowered as far as possible until all words fit
+  /// into a single line.
+  /// {@endtemplate}
+  final bool? wrapWords;
 
-  /// Defines the focus for this widget.
-  ///
-  /// Text is only selectable when widget is focused.
-  ///
-  /// The [focusNode] is a long-lived object that's typically managed by a
-  /// [StatefulWidget] parent. See [FocusNode] for more information.
-  ///
-  /// To give the focus to this widget, provide a [focusNode] and then
-  /// use the current [FocusScope] to request the focus:
-  ///
-  /// ```dart
-  /// FocusScope.of(context).requestFocus(myFocusNode);
-  /// ```
-  ///
-  /// This happens automatically when the widget is tapped.
-  ///
-  /// To be notified when the widget gains or loses the focus, add a listener
-  /// to the [focusNode]:
-  ///
-  /// ```dart
-  /// focusNode.addListener(() { print(myFocusNode.hasFocus); });
-  /// ```
-  ///
-  /// If null, this widget will create its own [FocusNode].
-  final FocusNode? focusNode;
+  /// {@template auto_size_text.overflowReplacement}
+  /// If the text is overflowing and does not fit its bounds, this widget is
+  /// displayed instead.
+  /// {@endtemplate}
+  final Widget? overflowReplacement;
 
-  /// {@macro flutter.widgets.editableText.onSelectionChanged}
-  final SelectionChangedCallback? onSelectionChanged;
+  /// {@template auto_size_text.onOverflow}
+  /// Called when the text overflows its container.
+  /// {@endtemplate}
+  final Function(bool overflow)? overflowCallback;
 
-  /// Called when the user taps on this selectable text.
+  /// Creates a [AutoSizeText]
   ///
-  /// The selectable text builds a [GestureDetector] to handle input events like tap,
-  /// to trigger focus requests, to move the caret, adjust the selection, etc.
-  /// Handling some of those events by wrapping the selectable text with a competing
-  /// GestureDetector is problematic.
-  ///
-  /// To unconditionally handle taps, without interfering with the selectable text's
-  /// internal gesture detector, provide this callback.
-  ///
-  /// To be notified when the text field gains or loses the focus, provide a
-  /// [focusNode] and add a listener to that.
-  ///
-  /// To listen to arbitrary pointer events without competing with the
-  /// selectable text's internal gesture detector, use a [Listener].
-  final GestureTapCallback? onTap;
+  /// If the [style] argument is null, the text will use the style from the
+  /// closest enclosing [DefaultTextStyle].
+  const AutoSizeText(
+    String this.data, {
+    Key? key,
+    this.textKey,
+    this.style,
+    this.strutStyle,
+    this.textAlign,
+    this.textDirection,
+    this.locale,
+    this.softWrap,
+    this.overflow,
+    this.textScaleFactor,
+    this.maxLines,
+    this.semanticsLabel,
+    this.textWidthBasis,
+    this.textHeightBehavior,
+    this.minFontSize,
+    this.maxFontSize,
+    this.stepGranularity,
+    this.presetFontSizes,
+    //this.group,
+    this.wrapWords,
+    this.overflowReplacement,
+    this.overflowCallback,
+  })  : textSpan = null,
+        super(key: key);
 
-  /// The ScrollPhysics to use when vertically scrolling the input.
-  ///
-  /// If not specified, it will behave according to the current platform.
-  final ScrollPhysics? scrollPhysics;
-
-  /// {@macro flutter.widgets.editableText.selectionControls}
-  final TextSelectionControls? selectionControls;
-
-  /// {@macro flutter.widgets.editableText.showCursor}
-  final bool showCursor;
-
-  /// Configuration of toolbar options.
-  ///
-  /// Paste and cut will be disabled regardless.
-  ///
-  /// If not set, select all and copy will be enabled by default.
-  final ToolbarOptions? toolbarOptions;
-
-  @override
-  _AutoSizeTextState createState() => _AutoSizeTextState();
-}
-
-class _AutoSizeTextState extends State<AutoSizeText> {
-  @override
-  void initState() {
-    super.initState();
-
-    widget.group?._register(this);
-  }
-
-  @override
-  void didUpdateWidget(AutoSizeText oldWidget) {
-    super.didUpdateWidget(oldWidget);
-
-    if (oldWidget.group != widget.group) {
-      oldWidget.group?._remove(this);
-      widget.group?._register(this);
-    }
-  }
+  /// Creates a [AutoSizeText] widget with a [TextSpan].
+  const AutoSizeText.rich(
+    TextSpan this.textSpan, {
+    Key? key,
+    this.textKey,
+    this.style,
+    this.strutStyle,
+    this.textAlign,
+    this.textDirection,
+    this.locale,
+    this.softWrap,
+    this.overflow,
+    this.textScaleFactor,
+    this.maxLines,
+    this.semanticsLabel,
+    this.textWidthBasis,
+    this.textHeightBehavior,
+    this.minFontSize,
+    this.maxFontSize,
+    this.stepGranularity,
+    this.presetFontSizes,
+    //this.group,
+    this.wrapWords,
+    this.overflowReplacement,
+    this.overflowCallback,
+  })  : data = null,
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(builder: (context, size) {
-      final defaultTextStyle = DefaultTextStyle.of(context);
-
-      var style = widget.style;
-      if (widget.style == null || widget.style!.inherit) {
-        style = defaultTextStyle.style.merge(widget.style);
-      }
-      if (style!.fontSize == null) {
-        style = style.copyWith(fontSize: AutoSizeText._defaultFontSize);
-      }
-
-      final maxLines = widget.maxLines ?? defaultTextStyle.maxLines;
-
-      _validateProperties(style, maxLines);
-
-      final result = _calculateFontSize(size, style, maxLines);
-      final fontSize = result[0] as double;
-      final textFits = result[1] as bool;
-
-      Widget text;
-
-      if (widget.group != null) {
-        widget.group!._updateFontSize(this, fontSize);
-        text = _buildText(widget.group!._fontSize, style, maxLines);
-      } else {
-        text = _buildText(fontSize, style, maxLines);
-      }
-
-      if (widget.overflowReplacement != null && !textFits) {
-        return widget.overflowReplacement!;
-      } else {
-        return text;
-      }
-    });
-  }
-
-  void _validateProperties(TextStyle style, int? maxLines) {
-    if (!widget._isSelectableText) {
-      assert(widget.overflow == null || widget.overflowReplacement == null,
-          'Either overflow or overflowReplacement must be null.');
-    }
-    assert(maxLines == null || maxLines > 0,
-        'MaxLines must be greater than or equal to 1.');
-    assert(widget.key == null || widget.key != widget.textKey,
-        'Key and textKey must not be equal.');
-
-    if (widget.presetFontSizes == null) {
-      assert(
-          widget.stepGranularity >= 0.1,
-          'StepGranularity must be greater than or equal to 0.1. It is not a '
-          'good idea to resize the font with a higher accuracy.');
-      assert(widget.minFontSize >= 0,
-          'MinFontSize must be greater than or equal to 0.');
-      assert(widget.maxFontSize > 0, 'MaxFontSize has to be greater than 0.');
-      assert(widget.minFontSize <= widget.maxFontSize,
-          'MinFontSize must be smaller or equal than maxFontSize.');
-      assert(widget.minFontSize / widget.stepGranularity % 1 == 0,
-          'MinFontSize must be a multiple of stepGranularity.');
-      if (widget.maxFontSize != double.infinity) {
-        assert(widget.maxFontSize / widget.stepGranularity % 1 == 0,
-            'MaxFontSize must be a multiple of stepGranularity.');
-      }
-    } else {
-      assert(widget.presetFontSizes!.isNotEmpty,
-          'PresetFontSizes must not be empty.');
-    }
-  }
-
-  List _calculateFontSize(
-      BoxConstraints size, TextStyle? style, int? maxLines) {
-    final span = TextSpan(
-      style: widget.textSpan?.style ?? style,
-      text: widget.textSpan?.text ?? widget.data,
-      children: widget.textSpan?.children,
-      recognizer: widget.textSpan?.recognizer,
-    );
-
-    final userScale =
-        widget.textScaleFactor ?? MediaQuery.textScaleFactorOf(context);
-
-    int left;
-    int right;
-
-    final presetFontSizes = widget.presetFontSizes?.reversed.toList();
-    if (presetFontSizes == null) {
-      final num defaultFontSize =
-          style!.fontSize!.clamp(widget.minFontSize, widget.maxFontSize);
-      final defaultScale = defaultFontSize * userScale / style.fontSize!;
-      if (_checkTextFits(span, defaultScale, maxLines, size)) {
-        return <Object>[defaultFontSize * userScale, true];
-      }
-
-      left = (widget.minFontSize / widget.stepGranularity).floor();
-      right = (defaultFontSize / widget.stepGranularity).ceil();
-    } else {
-      left = 0;
-      right = presetFontSizes.length - 1;
-    }
-
-    var lastValueFits = false;
-    while (left <= right) {
-      final mid = (left + (right - left) / 2).floor();
-      double scale;
-      if (presetFontSizes == null) {
-        scale = mid * userScale * widget.stepGranularity / style!.fontSize!;
-      } else {
-        scale = presetFontSizes[mid] * userScale / style!.fontSize!;
-      }
-      if (_checkTextFits(span, scale, maxLines, size)) {
-        left = mid + 1;
-        lastValueFits = true;
-      } else {
-        right = mid - 1;
-      }
-    }
-
-    if (!lastValueFits) {
-      right += 1;
-    }
-
-    double fontSize;
-    if (presetFontSizes == null) {
-      fontSize = right * userScale * widget.stepGranularity;
-    } else {
-      fontSize = presetFontSizes[right] * userScale;
-    }
-
-    return <Object>[fontSize, lastValueFits];
-  }
-
-  bool _checkTextFits(
-      TextSpan text, double scale, int? maxLines, BoxConstraints constraints) {
-    if (!widget.wrapWords) {
-      final words = text.toPlainText().split(RegExp('\\s+'));
-
-      final wordWrapTextPainter = TextPainter(
-        text: TextSpan(
-          style: text.style,
-          text: words.join('\n'),
-        ),
-        textAlign: widget.textAlign ?? TextAlign.left,
-        textDirection: widget.textDirection ?? TextDirection.ltr,
-        textScaleFactor: scale,
-        maxLines: words.length,
-        locale: widget.locale,
-        strutStyle: widget.strutStyle,
-      );
-
-      wordWrapTextPainter.layout(maxWidth: constraints.maxWidth);
-
-      if (wordWrapTextPainter.didExceedMaxLines ||
-          wordWrapTextPainter.width > constraints.maxWidth) {
-        return false;
-      }
-    }
-
-    final textPainter = TextPainter(
-      text: text,
-      textAlign: widget.textAlign ?? TextAlign.left,
-      textDirection: widget.textDirection ?? TextDirection.ltr,
-      textScaleFactor: scale,
-      maxLines: maxLines,
-      locale: widget.locale,
-      strutStyle: widget.strutStyle,
-    );
-
-    textPainter.layout(maxWidth: constraints.maxWidth);
-
-    return !(textPainter.didExceedMaxLines ||
-        textPainter.height > constraints.maxHeight ||
-        textPainter.width > constraints.maxWidth);
-  }
-
-  Widget _buildText(double fontSize, TextStyle style, int? maxLines) {
-    if (widget._isSelectableText) {
-      if (widget.data != null) {
-        return SelectableText(
-          widget.data!,
-          key: widget.textKey,
-          style: style.copyWith(fontSize: fontSize),
-          strutStyle: widget.strutStyle,
-          textAlign: widget.textAlign,
-          textDirection: widget.textDirection,
-          textScaleFactor: 1,
-          maxLines: maxLines,
-          autofocus: widget.autofocus,
-          cursorColor: widget.cursorColor,
-          cursorHeight: widget.cursorHeight,
-          cursorRadius: widget.cursorRadius,
-          cursorWidth: widget.cursorWidth,
-          dragStartBehavior: widget.dragStartBehavior,
-          enableInteractiveSelection: widget.enableInteractiveSelection,
-          focusNode: widget.focusNode,
-          onSelectionChanged: widget.onSelectionChanged,
-          onTap: widget.onTap,
-          scrollPhysics: widget.scrollPhysics,
-          selectionControls: widget.selectionControls,
-          showCursor: widget.showCursor,
-          toolbarOptions: widget.toolbarOptions,
-        );
-      } else {
-        return SelectableText.rich(
-          widget.textSpan!,
-          key: widget.textKey,
+    final span = textSpan ?? TextSpan(text: data);
+    return AutoSizeBuilder(
+      text: span,
+      style: style,
+      builder: (context, scale, overflow) {
+        overflowCallback?.call(overflow);
+        return Text.rich(
+          span,
+          key: textKey,
           style: style,
-          strutStyle: widget.strutStyle,
-          textAlign: widget.textAlign,
-          textDirection: widget.textDirection,
-          textScaleFactor: fontSize / style.fontSize!,
+          strutStyle: strutStyle,
+          textAlign: textAlign,
+          textDirection: textDirection,
+          locale: locale,
+          softWrap: softWrap,
+          overflow: this.overflow,
+          textScaleFactor: scale,
           maxLines: maxLines,
-          autofocus: widget.autofocus,
-          cursorColor: widget.cursorColor,
-          cursorHeight: widget.cursorHeight,
-          cursorRadius: widget.cursorRadius,
-          cursorWidth: widget.cursorWidth,
-          dragStartBehavior: widget.dragStartBehavior,
-          enableInteractiveSelection: widget.enableInteractiveSelection,
-          focusNode: widget.focusNode,
-          onSelectionChanged: widget.onSelectionChanged,
-          onTap: widget.onTap,
-          scrollPhysics: widget.scrollPhysics,
-          selectionControls: widget.selectionControls,
-          showCursor: widget.showCursor,
-          toolbarOptions: widget.toolbarOptions,
+          semanticsLabel: semanticsLabel,
         );
-      }
-    }
-
-    if (widget.data != null) {
-      return Text(
-        widget.data!,
-        key: widget.textKey,
-        style: style.copyWith(fontSize: fontSize),
-        strutStyle: widget.strutStyle,
-        textAlign: widget.textAlign,
-        textDirection: widget.textDirection,
-        locale: widget.locale,
-        softWrap: widget.softWrap,
-        overflow: widget.overflow,
-        textScaleFactor: 1,
-        maxLines: maxLines,
-        semanticsLabel: widget.semanticsLabel,
-      );
-    } else {
-      return Text.rich(
-        widget.textSpan!,
-        key: widget.textKey,
-        style: style,
-        strutStyle: widget.strutStyle,
-        textAlign: widget.textAlign,
-        textDirection: widget.textDirection,
-        locale: widget.locale,
-        softWrap: widget.softWrap,
-        overflow: widget.overflow,
-        textScaleFactor: fontSize / style.fontSize!,
-        maxLines: maxLines,
-        semanticsLabel: widget.semanticsLabel,
-      );
-    }
-  }
-
-  void _notifySync() {
-    setState(() {});
-  }
-
-  @override
-  void dispose() {
-    if (widget.group != null) {
-      widget.group!._remove(this);
-    }
-    super.dispose();
+      },
+      strutStyle: strutStyle,
+      minFontSize: minFontSize,
+      maxFontSize: maxFontSize,
+      stepGranularity: stepGranularity,
+      presetFontSizes: presetFontSizes,
+      textAlign: textAlign,
+      textDirection: textDirection,
+      locale: locale,
+      wrapWords: wrapWords,
+      overflowReplacement: overflowReplacement,
+      textScaleFactor: textScaleFactor,
+      maxLines: maxLines,
+      textWidthBasis: textWidthBasis,
+      textHeightBehavior: textHeightBehavior,
+    );
   }
 }
