@@ -2,12 +2,7 @@
 /// bounds.
 library auto_size_text;
 
-import 'dart:async';
-
-import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart' show SelectableText;
-import 'package:flutter/widgets.dart';
-
-part 'src/auto_size_text.dart';
-part 'src/auto_size_group.dart';
-part 'src/auto_size_group_builder.dart';
+export 'src/auto_size_builder/auto_size_builder.dart';
+export 'src/auto_size_text_field.dart';
+export 'src/auto_size_text.dart';
+export 'src/selectable_auto_size_text.dart';
