@@ -1,6 +1,5 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'utils.dart';
@@ -8,7 +7,7 @@ import 'utils.dart';
 void main() {
   testWidgets('Forces valid min and max fontSize', (tester) async {
     await tester.pumpWidget(
-      AutoSizeText(
+      const AutoSizeText(
         'AutoSizeText Test',
         style: TextStyle(fontSize: 25),
         minFontSize: -1,
@@ -17,7 +16,7 @@ void main() {
     expect(tester.takeException(), isAssertionError);
 
     await tester.pumpWidget(
-      AutoSizeText(
+      const AutoSizeText(
         'AutoSizeText Test',
         style: TextStyle(fontSize: 25),
         maxFontSize: 0,
@@ -26,7 +25,7 @@ void main() {
     expect(tester.takeException(), isAssertionError);
 
     await tester.pumpWidget(
-      AutoSizeText(
+      const AutoSizeText(
         'AutoSizeText Test',
         style: TextStyle(fontSize: 25),
         minFontSize: 20,
@@ -40,7 +39,7 @@ void main() {
     await pumpAndExpectFontSize(
       tester: tester,
       expectedFontSize: 15,
-      widget: SizedBox(
+      widget: const SizedBox(
         width: 10,
         height: 10,
         child: AutoSizeText(
@@ -56,7 +55,7 @@ void main() {
     await pumpAndExpectFontSize(
       tester: tester,
       expectedFontSize: 30,
-      widget: SizedBox(
+      widget: const SizedBox(
         width: 120,
         height: 40,
         child: AutoSizeText(
@@ -72,7 +71,7 @@ void main() {
     await pumpAndExpectFontSize(
       tester: tester,
       expectedFontSize: 20,
-      widget: DefaultTextStyle(
+      widget: const DefaultTextStyle(
         style: TextStyle(fontSize: 30),
         child: AutoSizeText(
           'AutoSizeText Test',
@@ -84,7 +83,7 @@ void main() {
     await pumpAndExpectFontSize(
       tester: tester,
       expectedFontSize: 20,
-      widget: AutoSizeText(
+      widget: const AutoSizeText(
         'AutoSizeText Test',
         style: TextStyle(fontSize: 30),
         maxFontSize: 20,
@@ -94,7 +93,7 @@ void main() {
     await pumpAndExpectFontSize(
       tester: tester,
       expectedFontSize: 20,
-      widget: AutoSizeText(
+      widget: const AutoSizeText(
         'AutoSizeText Test',
         style: TextStyle(fontSize: 20),
         maxFontSize: 30,
